Module aws_sdk_applicationinsights::error

Modules

access_denied_exception

See AccessDeniedException

bad_request_exception

See BadRequestException

internal_server_exception

See InternalServerException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

tags_already_exist_exception

See TagsAlreadyExistException

too_many_tags_exception

See TooManyTagsException

validation_exception

See ValidationException

Structs

AccessDeniedException

BadRequestException

The request is not understood by the server.

CreateApplicationError

CreateComponentError

CreateLogPatternError

DeleteApplicationError

DeleteComponentError

DeleteLogPatternError

DescribeApplicationError

DescribeComponentConfigurationError

DescribeComponentConfigurationRecommendationError

DescribeComponentError

DescribeLogPatternError

DescribeObservationError

DescribeProblemError

DescribeProblemObservationsError

InternalServerException

The server encountered an internal error and is unable to complete the request.

ListApplicationsError

ListComponentsError

ListConfigurationHistoryError

ListLogPatternSetsError

ListLogPatternsError

ListProblemsError

ListTagsForResourceError

ResourceInUseException

The resource is already created or in use.

ResourceNotFoundException

The resource does not exist in the customer account.

TagResourceError

TagsAlreadyExistException

Tags are already registered for the specified application ARN.

TooManyTagsException

The number of the provided tags is beyond the limit, or

UntagResourceError

UpdateApplicationError

UpdateComponentConfigurationError

UpdateComponentError

UpdateLogPatternError

ValidationException

The parameter is not valid.

Enums

CreateApplicationErrorKind

CreateComponentErrorKind

CreateLogPatternErrorKind

DeleteApplicationErrorKind

DeleteComponentErrorKind

DeleteLogPatternErrorKind

DescribeApplicationErrorKind

DescribeComponentConfigurationErrorKind

DescribeComponentConfigurationRecommendationErrorKind

DescribeComponentErrorKind

DescribeLogPatternErrorKind

DescribeObservationErrorKind

DescribeProblemErrorKind

DescribeProblemObservationsErrorKind

ListApplicationsErrorKind

ListComponentsErrorKind

ListConfigurationHistoryErrorKind

ListLogPatternSetsErrorKind

ListLogPatternsErrorKind

ListProblemsErrorKind

ListTagsForResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApplicationErrorKind

UpdateComponentConfigurationErrorKind

UpdateComponentErrorKind

UpdateLogPatternErrorKind