Module aws_sdk_applicationinsights::model

Modules

application_component

See ApplicationComponent

application_info

See ApplicationInfo

configuration_event

See ConfigurationEvent

log_pattern

See LogPattern

observation

See Observation

problem

See Problem

related_observations

See RelatedObservations

tag

See Tag

Structs

ApplicationComponent

Describes a standalone resource or similarly grouped resources that the application is made

ApplicationInfo

Describes the status of the application.

ConfigurationEvent

LogPattern

An object that defines the log patterns that belongs to a LogPatternSet.

Observation

Describes an anomaly or error with the application.

Problem

Describes a problem that is detected by correlating observations.

RelatedObservations

Describes observations related to the problem.

Tag

An object that defines the tags associated with an application.

Enums

CloudWatchEventSource

ConfigurationEventResourceType

ConfigurationEventStatus

FeedbackKey

FeedbackValue

LogFilter

OsType

SeverityLevel

Status

Tier