Struct aws_sdk_applicationinsights::model::observation::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Observation

Implementations

impl Builder

pub fn id(self, input: impl Into<String>) -> Self

The ID of the observation type.

pub fn set_id(self, input: Option<String>) -> Self

pub fn start_time(self, input: Instant) -> Self

The time when the observation was first detected, in epoch seconds.

pub fn set_start_time(self, input: Option<Instant>) -> Self

pub fn end_time(self, input: Instant) -> Self

The time when the observation ended, in epoch seconds.

pub fn set_end_time(self, input: Option<Instant>) -> Self

pub fn source_type(self, input: impl Into<String>) -> Self

The source type of the observation.

pub fn set_source_type(self, input: Option<String>) -> Self

pub fn source_arn(self, input: impl Into<String>) -> Self

The source resource ARN of the observation.

pub fn set_source_arn(self, input: Option<String>) -> Self

pub fn log_group(self, input: impl Into<String>) -> Self

The log group name.

pub fn set_log_group(self, input: Option<String>) -> Self

pub fn line_time(self, input: Instant) -> Self

The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

pub fn set_line_time(self, input: Option<Instant>) -> Self

pub fn log_text(self, input: impl Into<String>) -> Self

The log text of the observation.

pub fn set_log_text(self, input: Option<String>) -> Self

pub fn log_filter(self, input: LogFilter) -> Self

The log filter of the observation.

pub fn set_log_filter(self, input: Option<LogFilter>) -> Self

pub fn metric_namespace(self, input: impl Into<String>) -> Self

The namespace of the observation metric.

pub fn set_metric_namespace(self, input: Option<String>) -> Self

pub fn metric_name(self, input: impl Into<String>) -> Self

The name of the observation metric.

pub fn set_metric_name(self, input: Option<String>) -> Self

pub fn unit(self, input: impl Into<String>) -> Self

The unit of the source observation metric.

pub fn set_unit(self, input: Option<String>) -> Self

pub fn value(self, input: f64) -> Self

The value of the source observation metric.

pub fn set_value(self, input: Option<f64>) -> Self

pub fn cloud_watch_event_id(self, input: impl Into<String>) -> Self

The ID of the CloudWatch Event-based observation related to the detected problem.

pub fn set_cloud_watch_event_id(self, input: Option<String>) -> Self

pub fn cloud_watch_event_source(self, input: CloudWatchEventSource) -> Self

The source of the CloudWatch Event.

pub fn set_cloud_watch_event_source(
self,
input: Option<CloudWatchEventSource>
) -> Self

pub fn cloud_watch_event_detail_type(self, input: impl Into<String>) -> Self

The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

pub fn set_cloud_watch_event_detail_type(self, input: Option<String>) -> Self

pub fn health_event_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

pub fn set_health_event_arn(self, input: Option<String>) -> Self

pub fn health_service(self, input: impl Into<String>) -> Self

The service to which the AWS Health Event belongs, such as EC2.

pub fn set_health_service(self, input: Option<String>) -> Self

pub fn health_event_type_code(self, input: impl Into<String>) -> Self

The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

pub fn set_health_event_type_code(self, input: Option<String>) -> Self

pub fn health_event_type_category(self, input: impl Into<String>) -> Self

The category of the AWS Health event, such as issue.

pub fn set_health_event_type_category(self, input: Option<String>) -> Self

pub fn health_event_description(self, input: impl Into<String>) -> Self

The description of the AWS Health event provided by the service, such as Amazon EC2.

pub fn set_health_event_description(self, input: Option<String>) -> Self

pub fn code_deploy_deployment_id(self, input: impl Into<String>) -> Self

The deployment ID of the CodeDeploy-based observation related to the detected problem.

pub fn set_code_deploy_deployment_id(self, input: Option<String>) -> Self

pub fn code_deploy_deployment_group(self, input: impl Into<String>) -> Self

The deployment group to which the CodeDeploy deployment belongs.

pub fn set_code_deploy_deployment_group(self, input: Option<String>) -> Self

pub fn code_deploy_state(self, input: impl Into<String>) -> Self

The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

pub fn set_code_deploy_state(self, input: Option<String>) -> Self

pub fn code_deploy_application(self, input: impl Into<String>) -> Self

The CodeDeploy application to which the deployment belongs.

pub fn set_code_deploy_application(self, input: Option<String>) -> Self

pub fn code_deploy_instance_group_id(self, input: impl Into<String>) -> Self

The instance group to which the CodeDeploy instance belongs.

pub fn set_code_deploy_instance_group_id(self, input: Option<String>) -> Self

pub fn ec2_state(self, input: impl Into<String>) -> Self

The state of the instance, such as STOPPING or TERMINATING.

pub fn set_ec2_state(self, input: Option<String>) -> Self

pub fn rds_event_categories(self, input: impl Into<String>) -> Self

The category of an RDS event.

pub fn set_rds_event_categories(self, input: Option<String>) -> Self

pub fn rds_event_message(self, input: impl Into<String>) -> Self

The message of an RDS event.

pub fn set_rds_event_message(self, input: Option<String>) -> Self

pub fn s3_event_name(self, input: impl Into<String>) -> Self

The name of the S3 CloudWatch Event-based observation.

pub fn set_s3_event_name(self, input: Option<String>) -> Self

pub fn states_execution_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the step function execution-based observation.

pub fn set_states_execution_arn(self, input: Option<String>) -> Self

pub fn states_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the step function-based observation.

pub fn set_states_arn(self, input: Option<String>) -> Self

pub fn states_status(self, input: impl Into<String>) -> Self

The status of the step function-related observation.

pub fn set_states_status(self, input: Option<String>) -> Self

pub fn states_input(self, input: impl Into<String>) -> Self

The input to the step function-based observation.

pub fn set_states_input(self, input: Option<String>) -> Self

pub fn ebs_event(self, input: impl Into<String>) -> Self

The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

pub fn set_ebs_event(self, input: Option<String>) -> Self

pub fn ebs_result(self, input: impl Into<String>) -> Self

The result of an EBS CloudWatch event, such as failed or succeeded.

pub fn set_ebs_result(self, input: Option<String>) -> Self

pub fn ebs_cause(self, input: impl Into<String>) -> Self

The cause of an EBS CloudWatch event.

pub fn set_ebs_cause(self, input: Option<String>) -> Self

pub fn ebs_request_id(self, input: impl Into<String>) -> Self

The request ID of an EBS CloudWatch event.

pub fn set_ebs_request_id(self, input: Option<String>) -> Self

pub fn x_ray_fault_percent(self, input: i32) -> Self

The X-Ray request fault percentage for this node.

pub fn set_x_ray_fault_percent(self, input: Option<i32>) -> Self

pub fn x_ray_throttle_percent(self, input: i32) -> Self

The X-Ray request throttle percentage for this node.

pub fn set_x_ray_throttle_percent(self, input: Option<i32>) -> Self

pub fn x_ray_error_percent(self, input: i32) -> Self

The X-Ray request error percentage for this node.

pub fn set_x_ray_error_percent(self, input: Option<i32>) -> Self

pub fn x_ray_request_count(self, input: i32) -> Self

The X-Ray request count for this node.

pub fn set_x_ray_request_count(self, input: Option<i32>) -> Self

pub fn x_ray_request_average_latency(self, input: i64) -> Self

The X-Ray node request average latency for this node.

pub fn set_x_ray_request_average_latency(self, input: Option<i64>) -> Self

pub fn x_ray_node_name(self, input: impl Into<String>) -> Self

The name of the X-Ray node.

pub fn set_x_ray_node_name(self, input: Option<String>) -> Self

pub fn x_ray_node_type(self, input: impl Into<String>) -> Self

The type of the X-Ray node.

pub fn set_x_ray_node_type(self, input: Option<String>) -> Self

pub fn build(self) -> Observation

Consumes the builder and constructs a Observation