Struct aws_sdk_applicationinsights::model::Observation

#[non_exhaustive]
pub struct Observation {

    pub id: Option<String>,
    pub start_time: Option<Instant>,
    pub end_time: Option<Instant>,
    pub source_type: Option<String>,
    pub source_arn: Option<String>,
    pub log_group: Option<String>,
    pub line_time: Option<Instant>,
    pub log_text: Option<String>,
    pub log_filter: Option<LogFilter>,
    pub metric_namespace: Option<String>,
    pub metric_name: Option<String>,
    pub unit: Option<String>,
    pub value: Option<f64>,
    pub cloud_watch_event_id: Option<String>,
    pub cloud_watch_event_source: Option<CloudWatchEventSource>,
    pub cloud_watch_event_detail_type: Option<String>,
    pub health_event_arn: Option<String>,
    pub health_service: Option<String>,
    pub health_event_type_code: Option<String>,
    pub health_event_type_category: Option<String>,
    pub health_event_description: Option<String>,
    pub code_deploy_deployment_id: Option<String>,
    pub code_deploy_deployment_group: Option<String>,
    pub code_deploy_state: Option<String>,
    pub code_deploy_application: Option<String>,
    pub code_deploy_instance_group_id: Option<String>,
    pub ec2_state: Option<String>,
    pub rds_event_categories: Option<String>,
    pub rds_event_message: Option<String>,
    pub s3_event_name: Option<String>,
    pub states_execution_arn: Option<String>,
    pub states_arn: Option<String>,
    pub states_status: Option<String>,
    pub states_input: Option<String>,
    pub ebs_event: Option<String>,
    pub ebs_result: Option<String>,
    pub ebs_cause: Option<String>,
    pub ebs_request_id: Option<String>,
    pub x_ray_fault_percent: Option<i32>,
    pub x_ray_throttle_percent: Option<i32>,
    pub x_ray_error_percent: Option<i32>,
    pub x_ray_request_count: Option<i32>,
    pub x_ray_request_average_latency: Option<i64>,
    pub x_ray_node_name: Option<String>,
    pub x_ray_node_type: Option<String>,

}

Describes an anomaly or error with the application.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: Option<String>

The ID of the observation type.

start_time: Option<Instant>

The time when the observation was first detected, in epoch seconds.

end_time: Option<Instant>

The time when the observation ended, in epoch seconds.

source_type: Option<String>

The source type of the observation.

source_arn: Option<String>

The source resource ARN of the observation.

log_group: Option<String>

The log group name.

line_time: Option<Instant>

The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

log_text: Option<String>

The log text of the observation.

log_filter: Option<LogFilter>

The log filter of the observation.

metric_namespace: Option<String>

The namespace of the observation metric.

metric_name: Option<String>

The name of the observation metric.

unit: Option<String>

The unit of the source observation metric.

value: Option<f64>

The value of the source observation metric.

cloud_watch_event_id: Option<String>

The ID of the CloudWatch Event-based observation related to the detected problem.

cloud_watch_event_source: Option<CloudWatchEventSource>

The source of the CloudWatch Event.

cloud_watch_event_detail_type: Option<String>

The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

health_event_arn: Option<String>

The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

health_service: Option<String>

The service to which the AWS Health Event belongs, such as EC2.

health_event_type_code: Option<String>

The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

health_event_type_category: Option<String>

The category of the AWS Health event, such as issue.

health_event_description: Option<String>

The description of the AWS Health event provided by the service, such as Amazon EC2.

code_deploy_deployment_id: Option<String>

The deployment ID of the CodeDeploy-based observation related to the detected problem.

code_deploy_deployment_group: Option<String>

The deployment group to which the CodeDeploy deployment belongs.

code_deploy_state: Option<String>

The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

code_deploy_application: Option<String>

The CodeDeploy application to which the deployment belongs.

code_deploy_instance_group_id: Option<String>

The instance group to which the CodeDeploy instance belongs.

ec2_state: Option<String>

The state of the instance, such as STOPPING or TERMINATING.

rds_event_categories: Option<String>

The category of an RDS event.

rds_event_message: Option<String>

The message of an RDS event.

s3_event_name: Option<String>

The name of the S3 CloudWatch Event-based observation.

states_execution_arn: Option<String>

The Amazon Resource Name (ARN) of the step function execution-based observation.

states_arn: Option<String>

The Amazon Resource Name (ARN) of the step function-based observation.

states_status: Option<String>

The status of the step function-related observation.

states_input: Option<String>

The input to the step function-based observation.

ebs_event: Option<String>

The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

ebs_result: Option<String>

The result of an EBS CloudWatch event, such as failed or succeeded.

ebs_cause: Option<String>

The cause of an EBS CloudWatch event.

ebs_request_id: Option<String>

The request ID of an EBS CloudWatch event.

x_ray_fault_percent: Option<i32>

The X-Ray request fault percentage for this node.

x_ray_throttle_percent: Option<i32>

The X-Ray request throttle percentage for this node.

x_ray_error_percent: Option<i32>

The X-Ray request error percentage for this node.

x_ray_request_count: Option<i32>

The X-Ray request count for this node.

x_ray_request_average_latency: Option<i64>

The X-Ray node request average latency for this node.

x_ray_node_name: Option<String>

The name of the X-Ray node.

x_ray_node_type: Option<String>

The type of the X-Ray node.

Implementations

impl Observation

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Observation