Module aws_sdk_applicationinsights::operation

Structs

CreateApplication

Adds an application that is created from a resource group.

CreateComponent

Creates a custom component by grouping similar standalone instances to monitor.

CreateLogPattern

Adds an log pattern to a LogPatternSet.

DeleteApplication

Removes the specified application from monitoring. Does not delete the application.

DeleteComponent

Ungroups a custom component. When you ungroup custom components, all applicable

DeleteLogPattern

Removes the specified log pattern from a LogPatternSet.

DescribeApplication

Describes the application.

DescribeComponent

Describes a component and lists the resources that are grouped together in a component.

DescribeComponentConfiguration

Describes the monitoring configuration of the component.

DescribeComponentConfigurationRecommendation

Describes the recommended monitoring configuration of the component.

DescribeLogPattern

Describe a specific log pattern from a LogPatternSet.

DescribeObservation

Describes an anomaly or error with the application.

DescribeProblem

Describes an application problem.

DescribeProblemObservations

Describes the anomalies or errors associated with the problem.

ListApplications

Lists the IDs of the applications that you are monitoring.

ListComponents

Lists the auto-grouped, standalone, and custom components of the application.

ListConfigurationHistory

ListLogPatternSets

Lists the log pattern sets in the specific application.

ListLogPatterns

Lists the log patterns in the specific log LogPatternSet.

ListProblems

Lists the problems with your application.

ListTagsForResource

Retrieve a list of the tags (keys and values) that are associated with a specified

TagResource

Add one or more tags (keys and values) to a specified application. A

UntagResource

Remove one or more tags (keys and values) from a specified application.

UpdateApplication

Updates the application.

UpdateComponent

Updates the custom component name and/or the list of resources that make up the

UpdateComponentConfiguration

Updates the monitoring configurations for the component. The configuration input parameter

UpdateLogPattern

Adds a log pattern to a LogPatternSet.